correction according to review and fix clang-tidy

+++ tasks/mpi/kudryashova_i_gather my/func_tests/GatherMyMPIFuncTests.cpp
@@ -7,7 +7,7 @@ std::vector<int8_t> GetRandomVectorGatherMy(int size) {
   std::vector<int8_t> vector(size);
   std::srand(static_cast<unsigned>(time(nullptr)) + ++seedOffset);
   for (int i = 0; i < size; ++i) {
-    vector[i] = static_cast<int>(std::rand() % 201 - 100);
+    vector[i] = std::rand() % 201 - 100;
   }
   return vector;
 }
